requirements.txt
# ===== Core Django =====
django>=5.2.0
whitenoise>=6.0.0

# ===== Django REST Framework =====
djangorestframework>=3.15.0
djangorestframework-simplejwt>=5.3.0
drf-spectacular>=0.27.0
django-filter>=24.0

# ===== Database =====
dj-database-url>=2.1.0
psycopg2-binary>=2.9.9

# ===== Async Tasks =====
celery>=5.3.0
redis>=5.0.0
django-redis>=5.4.0

# ===== AI / Data Science =====
scikit-learn>=1.4.0
numpy>=1.26.0
scipy>=1.12.0

# ===== NLP / Sentiment =====
underthesea>=6.8.0

# ===== Utilities =====
Pillow>=10.0.0
matplotlib>=3.8.0
requests>=2.31.0
python-decouple>=3.8

# ===== Dev & Testing =====
pytest>=9.0.0
pytest-django>=4.8.0
